pyproject.toml
[build-system]
build-backend = "hatchling.build"
requires = [
  "hatchling",
]

[project]
name = "pytwitchify"
description = "Simple menu tool for launching Twitch streams."
readme = "README.md"
keywords = [
  "cli",
  "dmenu",
  "rofi",
  "twitch",
  "utility",
]
license = "MIT"
authors = [
  { name = "haaag", email = "git.haaag@gmail.com" },
]
requires-python = ">=3.8"
dynamic = [
  "version",
]
dependencies = [
  "pyselector>=0.0.6",
  "httpx==0.23.3",
  "rich==13.3.3",
  "python-dotenv==1.0.0",
]
[project.scripts]
pytwitchify = "pytwitchify.__main__:main"

[tool.hatch.version]
path = "src/pytwitchify/__about__.py"

[tool.black]
line-length = 120

[tool.pytest.ini_options]
pythonpath = "src"
testpaths = ["tests"]

[tool.coverage.report]
exclude_lines = [
  "no cov",
  "if __name__ == .__main__.:",
  "if TYPE_CHECKING:",
]

[tool.mypy]
show_error_codes = true
show_error_context = true
show_column_numbers = true
ignore_missing_imports = true
pretty = true
warn_no_return = false
warn_unused_ignores = true

[tool.ruff]
fix = false
line-length = 120
task-tags = ["HACK", "FIX", "BUG"]
show-source = true
show-fixes = true
target-version = "py37"
exclude = [
    ".venv",
    ".mypy_cache",
    ".pytest_cache",
]
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"
extend-select = [
    "B",    # (flake8-bugbear)
    "BLE",  # (flake8-blind-except)
    "C4",   # (flake8-comprehensions)
    "ERA",  # (eradicate)
    "I",    # (isort)
    "ICN",  # (flake8-import-conventions)
    "INP",  # (flake8-no-pep420)
    "ISC",  # (flake8-implicit-str-concat)
    "N",    # (pep8-naming)
    "NPY",  # (numpy-specific-rules)
    "PGH",  # (pygrep-hooks)
    "PIE",  # (flake8-pie)
    "PTH",  # (flake8-use-pathlib)
    "Q",    # (flake8-quotes)
    "RET",  # (flake8-return)
    "RSE",  # (flake8-raise)
    "RUF",  # (ruff-specific-rules)
    "S",    # (flake8-bandit)
    "SIM",  # (flake8-simplify)
    "T20",  # (flake8-print)
    "TCH",  # (flake8-type-checking)
    "TID",  # (flake8-tidy-imports)
    "YTT",  # (flake8-2020)
]

[tool.ruff.isort]
force-single-line = true
order-by-type = true
required-imports = ["from __future__ import annotations"]
